requirements.txt
requests==2.31.0
feedgen==1.0.0
python-dotenv==1.0.0
pydub==0.25.1
beautifulsoup4==4.12.2
schedule==1.2.0
google-cloud-texttospeech==2.16.3
openai==1.3.0
selenium==4.15.2
webdriver-manager==4.0.1